api/functions/requirements.txt
google-cloud-storage
google-cloud-pubsub
google-auth
firebase-admin
pympi-ling